broke all data processing into independent data directory
